+++ app/Models/Cliente.php
@@ -3,12 +3,13 @@
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use Intervention\Image\Facades\Image;
 
 class Cliente extends Model
 {
     protected $table = "clientes";
     protected $fillable = ['nome', 'imagem', 'link', 'descricao', 'status', 'home', 'paginas'];
-
+    
     public function newInfo($data)
     {
         $alias          = $this->getAlias($data['nome']);
@@ -26,10 +27,10 @@ class Cliente extends Model
             }
             $data['imagem'] = $this->uploadArquivo($data['arquivo'], $alias);
         }
-
+        
         $info = $this->update($data);
         return $info;
-    
+        
     }
     
     public function deleteInfo()
@@ -38,11 +39,20 @@ class Cliente extends Model
             $path = public_path('/storage/clientes/');
             $file = $this->imagem;
             $arquivo = $path.$file;
-            unlink($arquivo);
+            if(file_exists($arquivo)){
+                unlink($arquivo);
+            }
+            
+            $path2 = public_path('/storage/thumb/noticias/');
+            $arquivo2 = $path2.$file;
+            if(file_exists($arquivo2)){
+                unlink($arquivo2);
+            }
+            
         }
         
         $deleta = $this->delete();
-
+        
         return $deleta;
     }
     
@@ -54,6 +64,10 @@ class Cliente extends Model
         $path = public_path('/storage/clientes/');
         
         $file->move($path, $nomeArquivo);
+        
+        $image_resize = Image::make($path . $nomeArquivo);
+        $image_resize->resize(400, 250);
+        $image_resize->save(public_path('/storage/thumb/clientes/' . $nomeArquivo));
         
         if(!$file){
             flash('Falha ao fazer o upload do Arquivo')->warning();
@@ -70,6 +84,13 @@ class Cliente extends Model
         $path = public_path('/storage/clientes/');
         $file = $arquivo;
         $arquivo = $path.$file;
+        
+        $path2 = public_path('/storage/thumb/clientes/');
+        $arquivo2 = $path2.$file;
+        if(file_exists($arquivo2)){
+            unlink($arquivo2);
+        }
+        
         if(file_exists($arquivo)){
             unlink($arquivo);
             return true;
