+++ app/Models/Produto.php
@@ -3,6 +3,7 @@
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use Intervention\Image\Facades\Image;
 
 class Produto extends Model
 {
@@ -54,7 +55,16 @@ class Produto extends Model
             $path = public_path('/storage/produtos/');
             $file = $this->imagem;
             $arquivo = $path.$file;
-            unlink($arquivo);
+            if(file_exists($arquivo)){
+                unlink($arquivo);
+            }
+
+            $path2 = public_path('/storage/thumb/noticias/');
+            $arquivo2 = $path2.$file;
+            if(file_exists($arquivo2)){
+                unlink($arquivo2);
+            }
+
         }
         
         $deleta = $this->delete();
@@ -71,6 +81,10 @@ class Produto extends Model
         
         $file->move($path, $nomeArquivo);
         
+        $image_resize = Image::make($path . $nomeArquivo);
+        $image_resize->resize(400, 250);
+        $image_resize->save(public_path('/storage/thumb/produtos/' . $nomeArquivo));
+        
         if(!$file){
             flash('Falha ao fazer o upload do Arquivo')->warning();
             return redirect()
@@ -86,6 +100,13 @@ class Produto extends Model
         $path = public_path('/storage/produtos/');
         $file = $arquivo;
         $arquivo = $path.$file;
+
+        $path2 = public_path('/storage/thumb/produtos/');
+        $arquivo2 = $path2.$file;
+        if(file_exists($arquivo2)){
+            unlink($arquivo2);
+        }
+        
         if(file_exists($arquivo)){
             unlink($arquivo);
             return true;
@@ -101,7 +122,7 @@ class Produto extends Model
         $string = strtolower($string);
         return preg_replace(array("/(á|à|ã|â|ä)/","/(Á|À|Ã|Â|Ä)/","/(é|è|ê|ë)/","/(É|È|Ê|Ë)/","/(í|ì|î|ï)/","/(Í|Ì|Î|Ï)/","/(ó|ò|õ|ô|ö)/","/(Ó|Ò|Õ|Ô|Ö)/","/(ú|ù|û|ü)/","/(Ú|Ù|Û|Ü)/","/(ñ)/","/(Ñ)/", "/(ç|Ç)/"),explode(" ","a A e E i I o O u U n N c"),$string);    
     }
-
+    
     public function convertCommaToDot($valor){ 
         return str_replace(',','.', $valor);
     }
